Add source of new package imported in this new version of demo
In regard of the discussion in the issue #1171 (as in the comment (by eizamaliev)[https://github.com/opencv/open_model_zoo/issues/1171#issuecomment-637610746]) the package `monitors` is new package imported in this new version of the demo file. As it is a Cpp resource, it is probable for new users to mistaken it as a conventional PyPi package. Hence, the note helps them install it.